requirements.txt
discord.py[voice]==2.3.2
mafic==2.9.3
wavelink==3.0.0rc1
flask==2.3.3
humanize==4.8.0
asyncpraw==7.7.1
pymongo==4.5.0
requests==2.31.0
twine==4.0.2
urllib3==2.0.7
virtualenv==20.24.5
python-dotenv==1.0.0
dnspython==2.4.2
pillow==10.0.1
aiohttp==3.8.5
rich==13.5.2
pynacl==1.5.0
motor[srv]==3.3.1
beanie==1.22.6
pydantic==1.10.12
nest_asyncio==1.5.8
howlongtobeatpy==1.0.5
dynaconf==3.2.2
# pycord18n==1.0.3
git+https://github.com/YoungTrep/pycord18n.git
apscheduler==3.10.4
pytz==2023.3.post1
python-git-info==0.8.3
waitress==2.1.2